+++ CHANGELOG.md
@@ -2,12 +2,13 @@
 
 This changelog covers all three packages, as they are (for now) updated as a whole
 
-## Unreleased
+## v0.30.0
 
 - Add File management views. Preview images and videos, download them. #121
 - Add `uploadFiles` method to @tomic/lib. #121
 - Add upload field to forms #121
 - Fix bug resourcearray input #123
+- Add WebMonetization support #124
 
 ## v0.29.2